+++ packages/spritesheet/test/resources/spritesheet.json
@@ -1,0 +1,55 @@
+{
+    "frames": {
+        "pic-sensei.jpg": {
+            "frame": {
+                "x": 1,
+                "y": 1,
+                "w": 125,
+                "h": 125
+            },
+            "rotated": false,
+            "trimmed": false,
+            "spriteSourceSize": {
+                "x": 0,
+                "y": 0,
+                "w": 125,
+                "h": 125
+            },
+            "sourceSize": {
+                "w": 125,
+                "h": 125
+            }
+        },
+        "bunny.png": {
+            "frame": {
+                "x": 127,
+                "y": 1,
+                "w": 7,
+                "h": 10
+            },
+            "rotated": false,
+            "trimmed": false,
+            "spriteSourceSize": {
+                "x": 0,
+                "y": 0,
+                "w": 7,
+                "h": 10
+            },
+            "sourceSize": {
+                "w": 7,
+                "h": 10
+            }
+        }
+    },
+    "meta": {
+        "app": "http://github.com/odrick/free-tex-packer-core",
+        "version": "0.3.4",
+        "image": "spritesheet.png",
+        "format": "RGBA8888",
+        "size": {
+            "w": 134,
+            "h": 126
+        },
+        "scale": 1
+    }
+}
