+++ packages/spritesheet/test/resources/test.json
@@ -1,0 +1,4 @@
+{
+    "testNumber": 23,
+    "testString": "Test String 23"
+}
